Moved some of the interfaced from models to the encoder
