+++ jpegenc/interfaces/__init__.py
@@ -1,0 +1,8 @@
+
+from __future__ import absolute_import
+
+from .object_with_blocks import ObjectWithBlocks
+from .datastream import DataStream
+from .pixelstream import PixelStream
+from .rgbstream import RGBStream
+from .ycbcrstream import YCbCrStream
