+++ jpegenc/interfaces/object_with_blocks.py
