+++ jpegenc/models/__init__.py
@@ -1,7 +1,3 @@
 
-from .owb import ObjectWithBlocks
-
-from .interfaces import DataStream
-from .interfaces import RGBStream
 
 from .processing import ProcessingSubblock
